refactor: extract transform-urn-links-plugin
