+++ packages/gatsby-theme-api-docs/src/components/description.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import { visit } from 'unist-util-visit';
 import { designSystem, Markdown } from '@commercetools-docs/ui-kit';
 import { markdownFragmentToReact } from '@commercetools-docs/gatsby-theme-docs';
-import { useTypeLocations, locationForType } from '../hooks/use-type-locations';
-import { parseTypeURN } from '../utils/ctp-urn';
+import transformURNLinksPlugin from '../utils/transform-urn-links-plugin';
 
 const DescriptionParagraphContainer = styled.div`
   max-width: ${designSystem.dimensions.widths.pageContent};
@@ -14,25 +12,6 @@ const DescriptionParagraphContainer = styled.div`
 const DescriptionTextContainer = styled.span`
   display: inline-block;
 `;
-
-// a custom remark plugin that resolves URN style links from RAML descriptions to URLs
-const transformURNLinksPlugin = () => (ast) => {
-  const typeLocations = useTypeLocations();
-  visit(ast, 'link', (node) => {
-    const typeUrn = parseTypeURN(node.url);
-    if (typeUrn !== false) {
-      const typeUrl = locationForType(
-        typeUrn.apiKey,
-        typeUrn.name,
-        typeLocations
-      );
-      node.url =
-        typeUrl.url && typeof typeUrl.url === 'string'
-          ? typeUrl.url
-          : 'Content Error - type location not found for: ' + node.url;
-    }
-  });
-};
 
 // A pure description string that takes a markdown string
 // as a property. Use this to embed description into a layout like
